+++ torchtitan/experiments/rl/tests/integration_tests.py
@@ -37,6 +37,7 @@ def build_rl_test_list() -> list[OverrideDefinitions]:
                     "--trainer.parallelism.tensor_parallel_degree 2",
                     "--generator.parallelism.tensor_parallel_degree 2",
                     "--generator.sampling.n 2",
+                    "--batcher.batch.seq_len 256",
                     "--trainer.debug.no_batch_invariant",
                     "--generator.debug.no_batch_invariant",
                     "--compile.no-enable",
@@ -56,6 +57,7 @@ def build_rl_test_list() -> list[OverrideDefinitions]:
                     "--trainer.parallelism.tensor_parallel_degree 2",
                     "--generator.parallelism.tensor_parallel_degree 2",
                     "--generator.sampling.n 2",
+                    "--batcher.batch.seq_len 256",
                     "--trainer.debug.no_batch_invariant",
                     "--generator.debug.no_batch_invariant",
                     "--metrics.no-enable-wandb",
@@ -75,6 +77,7 @@ def build_rl_h100_test_list() -> list[OverrideDefinitions]:
                 [
                     "--module rl",
                     "--config rl_grpo_qwen3_0_6b_batch_invariant",
+                    "--batcher.batch.seq_len 256",
                     "--metrics.no-enable-wandb",
                 ],
             ],
